Change type of max to Long from Int

+++ src/main/scala/com/github/shokohara/googlepagination/GooglePagination.scala
@@ -22,14 +22,14 @@ object GooglePagination {
   def toPagination(v: (Boolean, (Int, Int, Int, Int, Int, Int, Int, Int, Int, Int), Boolean)): Pagination =
     toPagination(v._1 +: v._2 :+ v._3)
 
-  def maxPageNumber(max: Int, per: Int) = scala.math.ceil(max.toDouble / per.toDouble).toInt
-  def firstPaginationNumber(page: Int, max: Int, per: Int, width: Int): Int = {
+  def maxPageNumber(max: Long, per: Int) = scala.math.ceil(max.toDouble / per.toDouble).toInt
+  def firstPaginationNumber(page: Int, max: Long, per: Int, width: Int): Int = {
     val maxPage = this.maxPageNumber(max, per)
     if (page < 1 + width / 2) 1
     else if (maxPage - width / 2 < page) maxPage - width + 1
     else page - width / 2
   }
-  def pagination(page: Int, max: Int, per: Int, width: Int): Pagination = {
+  def pagination(page: Int, max: Long, per: Int, width: Int): Pagination = {
     val maxPage = this.maxPageNumber(max, per)
     val start = firstPaginationNumber(page, max, per, width)
     object limit extends Poly1 {
@@ -37,7 +37,7 @@ object GooglePagination {
     }
     (1 < page) +: sequenceTuple(0).map(limit) :+ (page < maxPage)
   }
-  def paginate(page: Int, max: Int, per: Int, width: Int): GooglePagination = {
+  def paginate(page: Int, max: Long, per: Int, width: Int): GooglePagination = {
     val p = pagination(page, max, per, width)
     val ps: List[Int] = p.tail.init.toList[Option[Int]].flatten
     GooglePagination(p.head, ps, p.last, maxPageNumber(max, per))
